django_api_task/requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
cffi==1.12.3
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
django-rest-knox==4.0.1
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
